src/test/java/com/example/springbootjpatuto1/service/MemberServiceTest.java
package com.example.springbootjpatuto1.service;

import com.example.springbootjpatuto1.domain.user.Member;
import com.example.springbootjpatuto1.repository.MemberRepository;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.annotation.Rollback;
import org.springframework.transaction.annotation.Transactional;

import static org.junit.jupiter.api.Assertions.*;

@SpringBootTest
@Transactional
class MemberServiceTest {

    @Autowired MemberService memberService;
    @Autowired MemberRepository memberRepository;

    @Test
    @Rollback(value = false)
    public void 회원가입() throws Exception {
        Member member = new Member();
        member.setName("kim");

        Long savedId = memberService.join(member);

        System.out.println("savedId = " + savedId);

        assertNotNull(savedId);
        
        Member findmember = memberService.findOne(savedId);

        System.out.println("findmember.getId() = " + findmember.getId());

        assertEquals(member, findmember);
    }

    @Test
    public void 중복_회원_예외() throws Exception {
        Member member1 = new Member();
        member1.setName("kim1");
        Member member2 = new Member();
        member2.setName("kim1");

        memberService.join(member1);
        assertThrows(IllegalStateException.class, () -> {
            memberService.join(member2);
        });
    }

}